requirements.txt
fastapi
uvicorn[standard]
livekit-api